bug on surface computing fixed; surface information not yet written in the headers.

+++ softwareTERAPIX/weightwatcher/src/prefs.h
@@ -54,7 +54,7 @@ typedef struct
   char		*oweight_name;				/* output WEIGHT */
   char		*oflag_name;				/* output FLAG */
   char          *getarea;                               /* compute area? */
-  float         weightlim;                              /* weight limit for area */
+  double         weightlim;                             /* weight limit for area */
   FLAGTYPE      geta_flags[MAXNIMAGE];                  /* flags for area */
   int           ngeta_flags;                            /* number of params */
   int		mem_bufsize;				/* # of buffer lines */
